requirements.txt
streamlit==1.39.0
wikipedia==1.4.0
langchain==0.3.14
google-generativeai==0.8.3
langchain_google==2.0.8
python-dotenv==1.0.1
PyPDF2==3.0.1
ipython==8.27.0
faiss-cpu==1.9.0.post1
langchain-community==0.3.14